Added simple user creation and authentication
Username and password has to be provided to be able to enter the chat. If a user with the provided username doesn't exist, the client will be able to send a user creation request to the server.
User data (users and passwords) are stored in memory and not in a file, so when the server restarts, it's lost.